+++ lib/angular2-tree-component.ts
@@ -1,5 +1,6 @@
 import { NgModule }      from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { Ng2MobxModule }      from 'ng2-mobx';
 
 import { TREE_ACTIONS, IActionMapping, IActionHandler } from './models/tree-options.model';
 import { ITreeOptions, IAllowDropFn } from './defs/api';
@@ -7,37 +8,40 @@ import { KEYS } from './constants/keys';
 import { TreeModel } from './models/tree.model';
 import { TreeNode } from './models/tree-node.model';
 import { TreeDraggedElement } from './models/tree-dragged-element.model';
+import { TreeVirtualScroll } from './models/tree-virtual-scroll.model';
 import { LoadingComponent } from './components/loading.component';
-import { LoadingComponent as DeprecatedLoadingComponent } from './components/deprecated-loading.component';
 import { TreeComponent } from './components/tree.component';
 import { TreeNodeComponent } from './components/tree-node.component';
 import { TreeNodeContent } from './components/tree-node-content.component';
-import { TreeNodeContent as DeprecatedTreeNodeContent } from './components/deprecated-tree-node-content.component';
 import { TreeNodeDropSlot } from './components/tree-node-drop-slot.component';
 import { TreeNodeExpanderComponent } from './components/tree-node-expander.component';
 import { TreeNodeChildrenComponent } from './components/tree-node-children.component';
+import { TreeNodeCollectionComponent } from './components/tree-node-collection.component';
+import { TreeViewportComponent } from './components/tree-viewport.component';
 import { TreeDropDirective } from './directives/tree-drop.directive';
 import { TreeDragDirective } from './directives/tree-drag.directive';
-import { AdHocComponentFactoryCreator } from './components/adhoc-component-factory.service';
 
 import './polyfills';
-import { deprecated } from './deprecated';
 
 const exportedDirectives = [
   TreeComponent,
   TreeNodeComponent,
   TreeNodeContent,
+  LoadingComponent,
   TreeDropDirective,
   TreeDragDirective,
   TreeNodeExpanderComponent,
   TreeNodeChildrenComponent,
-  TreeNodeDropSlot
+  TreeNodeDropSlot,
+  TreeNodeCollectionComponent,
+  TreeViewportComponent
 ];
 
 export {
   TreeModel,
   TreeNode,
   TreeDraggedElement,
+  TreeVirtualScroll,
   ITreeOptions,
   TREE_ACTIONS,
   KEYS,
@@ -52,12 +56,13 @@ export {
   TreeDragDirective,
   TreeNodeExpanderComponent,
   TreeNodeChildrenComponent,
-  TreeNodeDropSlot
+  TreeNodeDropSlot,
+  TreeNodeCollectionComponent,
+  TreeViewportComponent
 };
 
 @NgModule({
   declarations: [
-    LoadingComponent,
     ...exportedDirectives
   ],
   exports: [
@@ -65,33 +70,12 @@ export {
   ],
   imports: [
     CommonModule,
+    Ng2MobxModule
   ],
   providers: [
     TreeDraggedElement
   ]
 })
 export class TreeModule {}
-@NgModule({
-  declarations: [
-    DeprecatedLoadingComponent,
-    DeprecatedTreeNodeContent
-  ],
-  exports: [
-    TreeComponent,
-    TreeDropDirective,
-    TreeDragDirective
-  ],
-  imports: [
-    CommonModule,
-    TreeModule,
-  ],
-  providers: [
-    AdHocComponentFactoryCreator
-  ],
-})
-export class DeprecatedTreeModule {
-  constructor() {
-    deprecated('DeprecatedTreeModule', 'TreeModule for AoT compilation');
-  }
-}
+
 export default TreeModule;