+++ lib/components/loading.component.ts
@@ -1,15 +1,15 @@
-import { Component, Input, TemplateRef } from '@angular/core';
+import { Component, Input, TemplateRef, ElementRef } from '@angular/core';
 import { deprecatedSelector } from '../deprecated-selector';
 
 @Component({
-  selector: 'LoadingComponent, loading-component',
+  selector: 'LoadingComponent, tree-loading-component',
   template: `<span *ngIf="!template">loading...</span>
   <template [ngTemplateOutlet]="template"></template>`,
 })
 export class LoadingComponent {
   @Input() template: TemplateRef<any>;
 
-  constructor() {
-    deprecatedSelector('LoadingComponent', 'loading-component');
+  constructor(private elementRef: ElementRef) {
+    deprecatedSelector('LoadingComponent', 'tree-loading-component', elementRef);
   }
 }