+++ lib/components/tree-node-content.component.ts
@@ -1,4 +1,4 @@
-import { Component, Input, ViewEncapsulation, TemplateRef } from '@angular/core';
+import { Component, Input, ViewEncapsulation, TemplateRef, ElementRef } from '@angular/core';
 import { TreeNode } from '../models/tree-node.model';
 import { deprecatedSelector } from '../deprecated-selector';
 
@@ -16,7 +16,7 @@ export class TreeNodeContent {
   @Input() index: number;
   @Input() template: TemplateRef<any>;
 
-  constructor() {
-    deprecatedSelector('TreeNodeContent', 'tree-node-content');
+  constructor(private elementRef: ElementRef) {
+    deprecatedSelector('TreeNodeContent', 'tree-node-content', elementRef);
   }
 }