+++ lib/components/tree-node-drop-slot.component.ts
@@ -1,4 +1,4 @@
-import { Component, Input, ViewEncapsulation } from '@angular/core';
+import { Component, Input, ViewEncapsulation, ElementRef } from '@angular/core';
 import { TreeNode } from '../models/tree-node.model';
 import { deprecatedSelector } from '../deprecated-selector';
 
@@ -21,8 +21,8 @@ export class TreeNodeDropSlot {
   @Input() node: TreeNode;
   @Input() dropIndex: number;
 
-  constructor() {
-    deprecatedSelector('TreeNodeDropSlot', 'tree-node-drop-slot');
+  constructor(private elementRef: ElementRef) {
+    deprecatedSelector('TreeNodeDropSlot', 'tree-node-drop-slot', elementRef);
   }
 
   onDrop($event) {